backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
groq>=0.11.0
google-genai
langchain==0.2.16
langchain-groq==0.1.9
langchain-core==0.2.38
langchain-community==0.2.16
huggingface-hub>=0.20.0
transformers>=4.36.0
torch>=2.1.0
mediapipe>=0.10.9
opencv-python>=4.9.0
ultralytics>=8.1.0
numpy>=1.26.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
reportlab>=4.0.0
fpdf2>=2.7.0
pillow>=10.0.0
matplotlib>=3.8.0
bcrypt>=4.1.0
pydub>=0.25.0
deepface>=0.0.79
requests>=2.31.0
python-multipart>=0.0.6
werkzeug>=3.0.0
gunicorn>=21.0.0
